pyproject.toml
[tool.poetry]
name = "smart-traffic-management"
version = "0.1.0"
description = "This project uses YOLO for real-time vehicle detection and Fetch.AI agents to dynamically adjust traffic light timings based on vehicle density, optimizing traffic flow. It aims to reduce congestion and improve city traffic efficiency with decentralized AI control."
authors = ["YashGupta-24 <yashgupta.241103@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
uagents = "^0.19.0"
ultralytics = "^8.3.78"
opencv-python = "^4.11.0.86"
google-generativeai = "^0.8.4"
python-dotenv = "^1.0.1"
openai = "^1.64.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
